+++ packages/astro-tests/src/internal/log.ts
@@ -1,0 +1,7 @@
+import debugC from 'debug';
+
+export const debug = debugC('inox-tools:astro-tests');
+
+export function getDebug(name: string): debugC.Debugger {
+	return debug.extend(name);
+}