+++ packages/custom-routing/.eslintrc.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  rules: {
-    '@typescript-eslint/ban-types': 'off',
-    '@typescript-eslint/no-namespace': 'off',
-  },
-};