+++ packages/cut-short/src/internal/debug.ts
@@ -1,0 +1,7 @@
+import debugC from 'debug';
+
+export const debug = debugC('inox-tools:cut-short');
+
+export const getDebug = (segment?: string) => {
+	return segment ? debug.extend(segment) : debug;
+};