+++ packages/content-utils/CHANGELOG.md
@@ -1,5 +1,80 @@
 # @inox-tools/content-utils
 
+## 0.8.1
+
+### Patch Changes
+
+- 0dcdeaa: Bump Astro Integration Kit
+- Updated dependencies [0dcdeaa]
+  - @inox-tools/modular-station@0.4.1
+
+## 0.8.0
+
+### Minor Changes
+
+- Minimal bump to sync inter-dependencies to Astro 5 and Vite 6
+
+### Patch Changes
+
+- Updated dependencies
+  - @inox-tools/modular-station@0.4.0
+
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies [21e0744]
+- Updated dependencies [21e0744]
+  - @inox-tools/utils@0.3.0
+  - @inox-tools/modular-station@0.3.5
+
+## 0.7.4-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+  - @inox-tools/modular-station@0.3.4-beta.0
+
+## 0.7.4
+
+### Patch Changes
+
+- Updated dependencies [aa73961]
+- Updated dependencies [aa73961]
+  - @inox-tools/modular-station@0.3.4
+
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies [e07b8a8]
+  - @inox-tools/utils@0.2.0
+  - @inox-tools/modular-station@0.3.3
+
+## 0.7.2
+
+### Patch Changes
+
+- 8f04995: Add new Astro keywords
+- Updated dependencies [8f04995]
+  - @inox-tools/modular-station@0.3.2
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [b53264e]
+  - @inox-tools/utils@0.1.4
+  - @inox-tools/modular-station@0.3.1
+
+## 0.7.0
+
+### Minor Changes
+
+- 666543e: Allow retrieving the Git authors and co-authors of Content Collection Entries
+
 ## 0.6.1
 
 ### Patch Changes