+++ packages/inline-mod/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @inox-tools/inline-mod
 
+## 2.0.2
+
+### Patch Changes
+
+- 0dcdeaa: Bump Astro Integration Kit
+
+## 2.0.1
+
+### Patch Changes
+
+- Minimal bump to sync inter-dependencies to Astro 5 and Vite 6
+
+## 2.0.0
+
+### Major Changes
+
+- 21e0744: Bump support to Vite 6
+
+### Patch Changes
+
+- Updated dependencies [21e0744]
+- Updated dependencies [21e0744]
+  - @inox-tools/utils@0.3.0
+
+## 2.0.0-beta.0
+
+### Major Changes
+
+- b4843b9: Bump support to Vite 6
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+
+## 1.5.3
+
+### Patch Changes
+
+- aa73961: Remove global change to `debug` package options
+
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies [e07b8a8]
+  - @inox-tools/utils@0.2.0
+
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies [b53264e]
+  - @inox-tools/utils@0.1.4
+
 ## 1.5.0
 
 ### Minor Changes